feat/refactor: profile page redirects if user is not signed it. Created user information component
